refactor(config): split config.go into focused modules
- Split config.go into 3 focused files for better maintainability
- config.go: Core config  - ExtensionConfig, Config struct, load/save
- config_plugins.go: Plugin configs  - All plugin configuration types
- config_workspace.go: Workspace configs  - Discovery, modules
